Various fixes
Fixed a memory leak in Configuration.cpp
Made ProjectView now say the filename in the confirm file delete messagebox
Messed around with MapView->View->Persons As->Images a little bit more, but it turns out I'll have to rewrite it a bit perhaps...

since: entity_x = entity.x - 16; places it one tile left
but:   entity_x = entity.x - 8; doesn't move the image at all

I perhaps have to work out the offset's a little better.


git-svn-id: file:///desktop/SourceForge%20Backup/sphere/svnrepos/trunk@737 c0eb2ce6-7f40-480a-a1e0-6c4aea08c2c2

+++ sphere/source/editor/ProjectWindow.cpp
@@ -302,9 +302,6 @@ CProjectWindow::OnProjectItemOpen()
 afx_msg void
 CProjectWindow::OnProjectItemDelete()
 {
-  if (MessageBox("This will permanently delete the file.\nAre you sure you want to continue?", "Delete Project Item", MB_YESNO) == IDNO)
-    return;
-
   HTREEITEM item = m_TreeControl.GetSelectedItem();
   if (item == NULL)
     return;
@@ -334,6 +331,10 @@ CProjectWindow::OnProjectItemDelete()
   strcat(filename, m_Project->GetGroupDirectory(GroupType));
   strcat(filename, "\\");
   strcat(filename, m_TreeControl.GetItemText(item));
+
+  std::string message = "This will permanently delete the file \"" + std::string(filename) + "\".\nAre you sure you want to continue?";
+  if (MessageBox(message.c_str(), "Delete Project Item", MB_YESNO) == IDNO)
+    return;
 
   if (!DeleteFile(filename))
     MessageBox("Error: Could not delete file");
